+++ composer.json
@@ -20,9 +20,8 @@
       "PayWithCapture\\": "src",
       "PayWithCapture\\Services\\": "src/services",
       "PayWithCapture\\Builders\\": "src/builders",
-      "PayWithCapture\\Parsers\\": "src/parsers",
-      "PayWithCapture\\Contracts\\": "src/contracts",
-      "PayWithCapture\\Responses\\": "src/responses"
+      "PayWithCapture\\Validators\\": "src/validators",
+      "PayWithCapture\\Contracts\\": "src/contracts"
     }
   }
