--- src/responses/AccountPayment.php
@@ -1,9 +1,0 @@
-<?php
-namespace PayWithCapture\Responses;
-
-class AccountPayment
-{
-  public $orderId;
-  public $verify;
-  public $message;
-}
