--- src/responses/Authentication.php
@@ -1,10 +1,0 @@
-<?php
-namespace PayWithCapture\Responses;
-
-class Authentication
-{
-  public $tokenType;
-  public $accessToken;
-  public $expiresIn;
-  public $refreshToken;
-}
